database/stratos_scripts-harvesters/requirements.txt
mysql-connector-python==8.2.0

